backend/pyproject.toml
[project]
name = "spectra-backend"
version = "0.1.0"
description = "Spectra AI-powered data analytics backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pyjwt>=2.9.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.30.0",
    "aiofiles>=25.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "langsmith",
    "langchain-core>=0.3.0",
    "pyyaml>=6.0.0",
    "langchain-anthropic",
    "langchain-openai",
    "langchain-google-genai",
    "langchain-ollama>=0.2.0",
    "sse-starlette>=3.2.0",
    "e2b-code-interpreter>=1.0.2",
    "tavily-python>=0.7.0",
    "aiosmtplib>=3.0.0",
    "jinja2>=3.1.0",
    "apscheduler>=3.11.0,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
